html/semantics/embedded-content/the-video-element/video-loading-eager.tentative.html

<!DOCTYPE html>
<link rel="author" title="Squarespace" href="https://www.squarespace.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
  .below-viewport {
    position: absolute;
    top: 1000vh;
  }
</style>

<video id="below_viewport" class="below-viewport" src="/media/A4.mp4?pipe=trickle(d2)" loading="eager"></video>

<script>
  const belowViewportVideo = document.getElementById("below_viewport");

  async_test(t => {
    let loadeddataFired = false;

    belowViewportVideo.addEventListener("loadeddata", t.step_func(() => {
      loadeddataFired = true;
    }));

    window.addEventListener("load", t.step_func_done(() => {
      assert_true(
        loadeddataFired,
        "Below-viewport video with loading='eager' should fire loadeddata before window.load"
      );
    }));
  }, "Videos with loading='eager' load immediately regardless of viewport position and delay window load event");
</script>
